+++ frontend/src/App.css
@@ -21,7 +21,3 @@ main.app {
 .chat-output {
   flex-grow: 3;
 }
-
-.error {
-  color: red;
-}
