+++ frontend/src/components/ChatError/ChatError.css
@@ -1,0 +1,3 @@
+.chat-error {
+  color: red;
+}
