+++ frontend/src/components/ChatError/index.ts
@@ -1,0 +1,1 @@
+export { default as ChatError } from './ChatError';
