+++ frontend/src/components/ChatForm/types.ts
@@ -1,9 +1,3 @@
 export type TChatFormData = {
   prompt: string;
 };
-
-export interface IProps {
-  disable?: boolean;
-  onSubmit: (data: TChatFormData) => void;
-  onSetSystem: (data: TChatFormData) => void;
-}
